refactor: Phase 2.1 - Extract variable struct analysis into separate module
- Created impl/src/bitfield/variable/ directory structure
- Moved VariableStructAnalysis and analyze_variable_bits to variable/analysis.rs
- Implemented extension trait pattern for variable analysis
- Moved variable_bits_errors.rs to variable/errors.rs
- Updated imports to use the new module structure
- All tests passing (106/106)

This preserves ~90% of the original analyse.rs code by using extension traits.
