+++ impl/src/bitfield/variable/errors.rs
