+++ impl/src/bitfield/variable/expand.rs
@@ -1,0 +1,19 @@
+use crate::bitfield::{
+    config::Config,
+    BitfieldStruct,
+};
+use proc_macro2::TokenStream as TokenStream2;
+use quote::quote;
+
+/// Extension trait for variable struct expansion
+pub trait VariableStructExpander {
+    /// Expand variable-size struct
+    fn expand_variable_struct(&self, config: &Config) -> syn::Result<TokenStream2>;
+}
+
+impl VariableStructExpander for BitfieldStruct {
+    fn expand_variable_struct(&self, _config: &Config) -> syn::Result<TokenStream2> {
+        // TODO: Move variable expansion logic here in Phase 2.2
+        Ok(quote! {})
+    }
+}
